% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chValues.r
\name{ch.batchOverlap}
\alias{ch.batchOverlap}
\title{For each in a set of combinations, run ch.distOverlap Function}
\usage{
ch.batchOverlap(values, items, itemSet, df.combns, numRuns = 1000,
  outFile = NULL)
}
\arguments{
\item{values}{a vector of the values in the distributions.}

\item{items}{a vector that identifies the item that each value is linked to.}

\item{itemSet}{a vector of items that you want to run the bootstraps for.}

\item{df.combns}{a vector of combinations that identifies the items (by index number of itemSet) whose values should be combined and compared against every other combination in this vector.}

\item{numRuns}{the number of runs to do in the bootstrap. DEFAULT = 1000}

\item{outFile}{the name of a text file that will be output as the bootstrap runs.  This is filled incrementally just in case the program dies. DEFAULT = NULL - will not write a file}
}
\value{
a dataframe with the overlap statistics for each set of distributions compared. See ch.distOverlap
}
\description{
This function runs ch.distOverlap for all the combinations of distributions identified in df.combns.
}
\examples{
ch.batchOverlap (values= data1$responses, items = data1$items,itemSet = itemSet,df.combns = df.combns, numRuns=1000, outFile ="out.txt" )
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
