% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.batchOverlapFlex.r
\name{ch.batchOverlapFlex}
\alias{ch.batchOverlapFlex}
\title{For each in a set of combinations, run ch.distOverlapFlex Function}
\usage{
ch.batchOverlapFlex(
  values,
  items,
  itemSet,
  df.combns,
  numRuns = 1000,
  outFile = NULL,
  combFun = ch.maxAveComb,
  ...
)
}
\arguments{
\item{values}{a vector of the values in the distributions.}

\item{items}{a vector that identifies the item that each value is linked to.}

\item{itemSet}{a vector of items that you want to run the bootstraps for.}

\item{df.combns}{a vector of combinations that identifies the items (by index number of itemSet) whose values should be combined and compared against every other combination in this vector.}

\item{numRuns}{the number of runs to do in the bootstrap. DEFAULT = 1000.}

\item{outFile}{the name of a text file that will be output as the bootstrap runs.  This is filled incrementally just in case the program dies. DEFAULT = NULL - will not write a file}

\item{combFun}{If there are multiple items contributing to a single distribution, this function describes how the values will be combined across items in both the X and Y distributions. The function must combine elements of a list that might be of different different lengths. The default just flattens the list into one large vector.  DEFAULT = ch.maxAveComb  (with probMax = 0.5)}
}
\value{
a dataframe with the overlap statistics for each set of distributions compared. See ch.distOverlap
}
\description{
This function runs ch.distOverlap for all the combinations of distributions identified in df.combns.
}
\examples{
ch.batchOverlapFlex (data$responses, data$items,itemSet,df.combns,numRuns=1000, outFile ="out.txt" )
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
