% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chValues.r
\name{ch.combnVector}
\alias{ch.combnVector}
\title{New combinations Function}
\usage{
ch.combnVector(numItems, minNumPerSide = 1, maxNumPerSide = 1)
}
\arguments{
\item{numItems}{the total number of items used for the combinations.}

\item{minNumPerSide}{the minimum number of items in a combination. DEFAULT = 1.}

\item{maxNumPerSide}{the maxmimun number of items in a combination. DEFAULT = 1.}
}
\value{
a vector of all the combinations (by item number)
}
\description{
This function creates a vector of all the combinations of itemList taken minNumPerSide at a time to maxNumPerSide at a time.
}
\examples{
ch.combnVector (5, 1,3)
}
\keyword{combinations}
\keyword{values}
