% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chValues.r
\name{ch.distOverlap}
\alias{ch.distOverlap}
\title{Bootstrap the overlap of two distributions Function}
\usage{
ch.distOverlap(xDist, yDist, numRuns)
}
\arguments{
\item{xDist}{Distribution X.}

\item{yDist}{Distribution Y.}

\item{numRuns}{the number of runs in the bootstrap. DEFAULT = 100.}
}
\value{
the average percentage of samples that xDist was greater than yDist, the sd of the average percentage, Overlap = a transformation of average percentage so 1=complete overlap and 0 = no overlap, direction = 1 if xDist > yDist; else -1
}
\description{
This function does a bootstrap to calculate the percentage of samples that Distribution X is greater than Distribution Y.
}
\examples{
ch.distOverlap (xValue, yValue,100)
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
