% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.distOverlapMaxAveCombXY.r
\name{ch.distOverlapMaxAveCombXY}
\alias{ch.distOverlapMaxAveCombXY}
\title{Bootstrap the overlap of two distributions Function with a flexible function}
\usage{
ch.distOverlapMaxAveCombXY(
  xDist,
  yDist,
  numRuns,
  probMaxX = 0.5,
  probMaxY = 0.5
)
}
\arguments{
\item{xDist}{a list containing the values for Distribution X.}

\item{yDist}{a list containing the values for Distribution Y.}

\item{numRuns}{the number of runs in the bootstrap. DEFAULT = 100.}

\item{probMaxX}{A probability that specifies the probability of samples that should be devoted to the xDist list column with the max(median()). When probMax = 0, the column with the max(median()) has no influence on either the average or above and beyond the average.  When probMax = 0.5, the column with the max(median()) is 50 percent of the output vector.  This is simply collapsing into a vector (as.vector) when there are only 2 vectors in the list. But it will have an outsized influence if there are more than two vector in the list.  When probMax = 1, the column with the max(median()) is 100 percent of the output vector.  DEFAULT = 0.5 (equal influence for two vectors)(probMax = .8 will produce a distribution in which .4 of the total are all the items (including the max) and .6 of the total are MAX items - when there are only two vectors in the list). Use ch.getProbMax() to calculate the appropriate probMax.}

\item{probMaxY}{The same as probMaxX, but for the yDist list.}
}
\value{
the average percentage of samples that xDist was greater than yDist, the sd of the average percentage, Overlap = a transformation of average percentage so 1=complete overlap and 0 = no overlap, direction = 1 if xDist > yDist; else -1
}
\description{
This function does a bootstrap to calculate the percentage of samples that Distribution X is greater than Distribution Y. When Distribution X (or Y) are the combination of multiple items, then it uses combFun (a function) to combine the values of those items into a single distribution.
}
\examples{
ch.distOverlapMaxAveCombXY (xValue, yValue,100, combFun = ch.maxAveComb, probMax = 0.667)
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
