% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.filterValuesData.r
\name{ch.filterValuesData}
\alias{ch.filterValuesData}
\title{A function to filter the values data for analysis}
\usage{
ch.filterValuesData(data, snCol, promptCol, valueCol, RTcol, params)
}
\arguments{
\item{data}{the values dataframe.}

\item{snCol}{a string that specifies the name of the column in "data" that contains the subject number.}

\item{promptCol}{a string that specifies the name of the column in "data" that contains the prompt for each trial.}

\item{valueCol}{a string that specifies the name of the column in "data" that contains the participant's estimate of value for the item in each trial.}

\item{params}{a list of parameters that are read in using "ch.readValuesDBfile.r."}
}
\value{
a dataframe of filtered data.  It also writes the items; the item statistics; and the data; to the files specified in params.
}
\description{
This function filters the values data for analysis by removing RT outliers and bad subjects based on RT. The thresholds for removal are set in the valuessDBfile that is read and stored in a parameter list.
}
\examples{
ch.filterValuesData (data=valuesData, "sn", "prompt", "resp", params=parameters)
}
\keyword{data}
\keyword{filter}
\keyword{values}
