% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getItemsSetsFromExperimentFile.r
\name{ch.getItemSetsFromChoiceDataframe}
\alias{ch.getItemSetsFromChoiceDataframe}
\title{A function to extract a the items present in an experiment file}
\usage{
ch.getItemSetsFromChoiceDataframe(
  choiceDataframe = NULL,
  outputFile = NULL,
  itemColumns = c("IA1", "IB1"),
  practiceColumn = NULL,
  practiceValue = "practice"
)
}
\arguments{
\item{choiceDataframe}{A dataframe with data from the experiment.}

\item{outputFile}{the filename that you want the output written to. DEFAULT = NULL (no file written)}

\item{practiceColumn}{the name of the column that contains the values indicating whether or not the trials are practice. DEFAULT = NULL (if NULL, no practice filtering will occur)}

\item{practiceValue}{the value in practiceColumn that specifies the trial is a practice trial. DEFAULT = "practice"}

\item{itemColumn}{the name of the column that contains the item values. DEFAULT = c("IA1", "IB1")}
}
\value{
a vector with the unique items used in the experiment
}
\description{
A function to extract a the items present in an experiment file
}
\examples{
ch.getItemSetsFromExperimentFile (experiment.txt,output.txt, probes, practice, 0)
}
\keyword{extract}
\keyword{file}
\keyword{items}
\keyword{morals}
