% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getItemsFromExperimentFile.r
\name{ch.getItemsFromExperimentFile}
\alias{ch.getItemsFromExperimentFile}
\title{A function to extract a the items present in an experiment file}
\usage{
ch.getItemsFromExperimentFile(experimentFile = NULL, outputFile = NULL,
  itemColumn = "ItemX", practiceColumn = "typeOfScen",
  practiceValue = "practice")
}
\arguments{
\item{experimentFile}{a tab delimited file with data from the experiment.}

\item{outputFile}{the filename that you want the output written to. DEFAULT = NULL (no file written)}

\item{itemColumn}{the name of the column that contains the item values. DEFAULT = "ItemX"}

\item{practiceColumn}{the name of the column that contains the values indicating whether or not the trials are practice. DEFAULT = "typeOfScen"}

\item{practiceValue}{the value in practiceColumn that specifies the trial is a practice trial. DEFAULT = "practice"}
}
\value{
a vector with the unique items used in the experiment
}
\description{
A function to extract a the items present in an experiment file
}
\examples{
ch.getItemsFromExperimentFile (experiment.txt,output.txt, probes, practice, 0)
}
\keyword{extract}
\keyword{file}
\keyword{items}
\keyword{morals}
