% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getProbeValues.r
\name{ch.getProbeValues}
\alias{ch.getProbeValues}
\title{A function to extract a the raw value data for a set of probes.}
\usage{
ch.getProbeValues(data, probeItems, promptColumn = "prompt", outputFile = NULL)
}
\arguments{
\item{data}{a datafrane containing the raw value data from an experiment or set of experiments.}

\item{probeItems}{A vector of probe items that you want the value data to be extracted from "data"}

\item{promptColumn}{A string specifying the name of the column in "data" that contains the prompt names. DEFAULT = "prompt"}

\item{outputFile}{the filename that you want the output written to. DEFAULT = NULL (no file written)}
}
\value{
a dataframe containing the raw values and associated information for the prompts in promptItems
}
\description{
A function to extract a the raw value data for a set of probes.
}
\examples{
ch.getProbeValues (df.myInputData, c("a nun", "your mother"), outputFile = "out.txt")
}
\keyword{extract}
\keyword{file}
\keyword{prompt}
\keyword{value}
