% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.getValuesSummaryStats.r
\name{ch.getValuesSummaryStats}
\alias{ch.getValuesSummaryStats}
\title{A function to get the value stats for each item}
\usage{
ch.getValuesSummaryStats(data, promptCol, valueCol, RTcol, outFile = NULL)
}
\arguments{
\item{data}{the values dataframe.}

\item{promptCol}{a string that specifies the name of the column in "data" that contains the prompt for each trial.}

\item{valueCol}{a string that specifies the name of the column in "data" that contains the participant's estimate of value for the item in each trial.}

\item{RTcol}{a string that specifies the name of the column in "data" that contains the participant's reaction times for the item in each trial.}

\item{outFile}{a string that specifies the name of the file to output the stats in. DEFAULT = NULL (no output)}
}
\value{
a dataframe of the value stats.  It also writes the item statistics to the outFile.
}
\description{
This function gets the value stats for each item and outputs those into a file.
}
\examples{
ch.getValuesSummaryStats (data=valuesData, "prompt", "respS", "respTime")
}
\keyword{data}
\keyword{filter}
\keyword{values}
