% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.listMultiElementComb.r
\name{ch.listMultiElementComb}
\alias{ch.listMultiElementComb}
\title{This function combines a list into a single vector}
\usage{
ch.listMultiElementComb(list.data, combType = "mean", replace = TRUE)
}
\arguments{
\item{combType}{A string that specifies the combination statistic.  Currently, two are implemented: "mean" and "sum"}

\item{replace}{A boolean that specifies whether the sample is take with replacement or not. DEFAULT = TRUE}

\item{df.data}{A list with each element filled with numbers.}
}
\value{
a vector of values that combines all the columns in df.data with probMax samples devoted to the column with the max(median())
}
\description{
This function combines a list into a single vector by applying a function across the elements. Essentially, this function will identify the minimum length of all the elements (minN), then sample each element with a size of minN. It will treat each sample as a column in a dataframe and apply rowMeans or rowSums.
}
\examples{
ch.listMultiElementComb (df.data, combType = "mean")
}
\keyword{combine}
\keyword{function}
\keyword{vectors}
