% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.maxAveComb.r
\name{ch.maxAveComb}
\alias{ch.maxAveComb}
\title{This function combines a list into a single vector}
\usage{
ch.maxAveComb(list.data, probMax = 0.5)
}
\arguments{
\item{probMax}{A probability that specifies the probability of samples that should be devoted to the column with the max(median()). DEFAULT = 0.5 (so it is simply a collapsing into a vector (as.vector)); (probMax = .8 will produce a distribution in which .4 of the total are all the items (including the max) and .6 of the total are MAX items.)}

\item{df.data}{A list with each element filled with numbers.}
}
\value{
a vector of values that combines all the columns in df.data with probMax samples devoted to the column with the max(median())
}
\description{
This function combines a list into a single vector controling the percent of observations devoted to the list element with the max median.
}
\examples{
ch.maxAveComb (df.data, probMax = .7)
}
\keyword{combine}
\keyword{function}
\keyword{vectors}
