% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.MCbatchOverlapPromptFile.r
\name{ch.MCbatchOverlapPromptFile}
\alias{ch.MCbatchOverlapPromptFile}
\title{Calculate the distributional overlaps for item combinations specified in the df.prompts dataframe, multicore version}
\usage{
ch.MCbatchOverlapPromptFile(
  values,
  items,
  df.prompts,
  itemAcolNames = "IA",
  itemBcolNames = "IB",
  numRuns = 1000,
  outFile = NULL,
  multicore = TRUE,
  verbose = FALSE,
  combFun = ch.maxAveComb,
  ...
)
}
\arguments{
\item{values}{A vector of the values in the distributions for each item.}

\item{items}{A vector that identifies the item that each value (in the values option) is linked to.}

\item{df.prompts}{A dataframe of item combinations that you want to run the bootstraps for. The each column should specify one item of the combination to be compared. Use "NA" for items in a column whereby no item is present.}

\item{itemAcolNames}{A vector of strings that identifies the column names in df.prompts that contain the items whose values should be combined and compared to those items identified in the itemBcolNames columns.}

\item{itemBcolNames}{A vector of strings that identifies the column names in df.prompts that contain the items whose values should be combined and compared to those items identified in the itemAcolNames columns.}

\item{numRuns}{the number of runs to do in the bootstrap. DEFAULT = 1000.}

\item{outFile}{the name of a text file that will be output as the bootstrap runs.  This is filled incrementally just in case the program dies. DEFAULT = NULL - will not write a file}

\item{multicore}{A boolean specifying whether the bootstrap runs in parallel using multiple cores. DEFAULT = TRUE}

\item{verbose}{A boolean specifying whether to print the progress of the bootstrap. DEFAULT = FALSE}

\item{combFun}{If there are multiple items contributing to a single distribution, this function describes how the values will be combined across items in both the X and Y distributions. The function must combine elements of a list that might be of different different lengths. The default just flattens the list into one large vector.  DEFAULT = ch.maxAveComb  (with probMax = 0.5)}
}
\value{
a dataframe with the overlap statistics for each set of distributions compared. See ch.distOverlap
}
\description{
This function runs ch.distOverlap for all the combinations of distributions identified in df.combns.
}
\examples{
ch.batchOverlapPromptFile (data$responses, data$items,itemSet,df.prompts, c("IA1", "IA2"), c("IB1"), 1000, outFile ="out.txt" )
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
