% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.overlapToP.r
\name{ch.overlapToP}
\alias{ch.overlapToP}
\title{This function converts Overlap to the p(D1 > D2) .}
\usage{
ch.overlapToP(overlap)
}
\arguments{
\item{overlap}{A number that equals 1 - [abs (p(D1 > D2) - 0.5) / 0.5].}
}
\value{
p(D1 > D2)
}
\description{
Assume two overlapping distributions, D1 and D2. Randomly select a value from D1 and D2.  p(D1 > D2) is the probability that you select a value from D1 that is greater than D2.  Overlap = 1 - [abs (p(D1 > D2) - 0.5) / 0.5].  This function converts Overlap back to p(D1 > D2).
}
\examples{
ch.overlapToP (0.5)
}
\keyword{>}
\keyword{D2)}
\keyword{meanP}
\keyword{overlap}
\keyword{p(D1}
