% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.pToDirOv.r
\name{ch.pToDirOv}
\alias{ch.pToDirOv}
\title{This function converts Overlap and direction to a directional overlap.}
\usage{
ch.pToDirOv(pIn, dirIn)
}
\arguments{
\item{pIn}{When randomly sampling from two distributions (D1 and D2), this is the p(D1 > D2) .}

\item{direction}{A number that equals -1 or 1. The direction specifies the whether D1 or D2 was the greater distribution with the greater value. It is output when p(D1 > D2) is calculated with the bootstrap.}
}
\value{
Directional Overlap
}
\description{
Assume two overlapping distributions, D1 and D2. Randomly select a value from D1 and D2.  p(D1 > D2) is the probability that you select a value from D1 that is greater than D2.  Overlap = 1 - [abs (p(D1 > D2) - 0.5) / 0.5].  Directional Overlap = abs(overlap - 1) * direction. Directional overlap goes from -1 (no overlap D1 smaller than D2) to 0 (D1 == D2) to 1 (no overlap D1 larger than D2)
}
\examples{
ch.pToDirOv (0.5, 1)
}
\keyword{>}
\keyword{D2)}
\keyword{directional}
\keyword{meanP}
\keyword{overlap}
\keyword{p(D1}
