% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.pToOverlap.r
\name{ch.pToOverlap}
\alias{ch.pToOverlap}
\title{This function converts p(D1 > D2) to Overlap.}
\usage{
ch.pToOverlap(pIn)
}
\arguments{
\item{pIn}{When randomly sampling from two distributions (D1 and D2), this is the p(D1 > D2) .}
}
\value{
Overlap
}
\description{
Assume two overlapping distributions, D1 and D2. Randomly select a value from D1 and D2.  p(D1 > D2) is the probability that you select a value from D1 that is greater than D2.  Overlap = 1 - [abs (p(D1 > D2) - 0.5) / 0.5].  This function converts p(D1 > D2) to Overlap.
}
\examples{
ch.pToOverlap (0.5)
}
\keyword{>}
\keyword{D2)}
\keyword{meanP}
\keyword{overlap}
\keyword{p(D1}
