% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.readValuessDBfile.r
\name{ch.readValuessDBfile}
\alias{ch.readValuessDBfile}
\title{A function to read the values dbfile}
\usage{
ch.readValuessDBfile(filename = "valuesDBfile.txt")
}
\arguments{
\item{dbfile}{the filename of the dbfile.}
}
\value{
a list with all the parameters
}
\description{
This function reads the dbfile for the values analysis.  The dbfile contains key words and their values for the analysis.
}
\examples{
ch.readMoralsDBfile ("myDBfile.txt")
}
\keyword{dbfile}
\keyword{read}
\keyword{values}
