% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.valuesBootstrapOverlaps.r
\name{ch.valuesBootstrapOverlaps}
\alias{ch.valuesBootstrapOverlaps}
\title{A function to calculate the overlaps of the values data using a bootstrap}
\usage{
ch.valuesBootstrapOverlaps(
  data,
  valuesCol,
  promptCol,
  params,
  allItems = T,
  multicore = F,
  combFun = ch.maxAveComb,
  ...
)
}
\arguments{
\item{data}{the values dataframe.}

\item{valuesCol}{a string that specifies the name of the column in "data" that contains the participant's estimate of value for the item in each trial.}

\item{promptCol}{a string that specifies the name of the column in "data" that contains the prompt for each trial.}

\item{params}{a list of parameters that are read in using "ch.readValuesDBfile.r."}

\item{allItems}{a boolean to state whether all the items are to be bootsrapped.  If FALSE, then the "itemSetDataFile" in the params list must contain a subset of items to be bootstrapped. DEFAULT = TRUE.}

\item{multicore}{a boolean to specifies whether to run the analysis in multicore mode. DEFAULT = FALSE.}

\item{combFun}{If there are multiple items contributing to a single distribution, this function describes how the values will be combined across items in both the X and Y distributions. The function must combine elements of a list that might be of different different lengths. The default just flattens the list into one large vector.  DEFAULT = ch.maxAveComb  (with probMax = 0.5)}
}
\value{
a dataframe of the overlaps by probe combination.  It also writes the data to the file (bootstrapOutdataFile) specified in params.
}
\description{
This function calculate the overlaps of the values data using a bootstrap.
}
\examples{
ch.valuesBootstrapOverlaps (data=valuesData, "tValues", "prompt", params=parameters)
}
