% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.batchOverlapFlex.r
\name{ch.batchOverlapFlex}
\alias{ch.batchOverlapFlex}
\title{For each in a set of combinations, run ch.distOverlapFlex Function}
\usage{
ch.batchOverlapFlex(values, items, itemSet, df.combns, numRuns,
  combFun = as.vector, outFile = NULL)
}
\arguments{
\item{values}{a vector of the values in the distributions.}

\item{items}{a vector that identifies the item that each value is linked to.}

\item{itemSet}{a vector of items that you want to run the bootstraps for.}

\item{df.combns}{a vector of combinations that identifies the items (by index number of itemSet) whose values should be combined and compared against every other combination in this vector.}

\item{numRuns}{the number of runs to do in the bootstrap.}

\item{combFun}{If there are multiple items contributing to a single distribution, this function describes how the values will be combined across items. The function must combine the rows of a matrix and return the result as a vector (e.g., rowMeans). The default just flattens the matrix into one large vector. DEFAULT = as.vector}

\item{outFile}{the name of a text file that will be output as the bootstrap runs.  This is filled incrementally just in case the program dies. DEFAULT = NULL - will not write a file}
}
\value{
a dataframe with the overlap statistics for each set of distributions compared. See ch.distOverlap
}
\description{
This function runs ch.distOverlap for all the combinations of distributions identified in df.combns.
}
\examples{
ch.batchOverlap (data$responses, data$items,itemSet,df.combns, "out.txt" )
}
\keyword{bootstrap}
\keyword{distribution}
\keyword{overlap}
\keyword{values}
