% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.valuesBootstrapOverlaps.r
\name{ch.valuesBootstrapOverlaps}
\alias{ch.valuesBootstrapOverlaps}
\title{A function to calculate the overlaps of the values data using a bootstrap}
\usage{
ch.valuesBootstrapOverlaps(data, valuesCol, promptCol, params)
}
\arguments{
\item{data}{the values dataframe.}

\item{promptCol}{a string that specifies the name of the column in "data" that contains the prompt for each trial.}

\item{params}{a list of parameters that are read in using "ch.readValuesDBfile.r."}

\item{valueCol}{a string that specifies the name of the column in "data" that contains the participant's estimate of value for the item in each trial.}
}
\value{
a dataframe of the overlaps by probe combination.  It also writes the data to the file (bootstrapOutdataFile) specified in params.
}
\description{
This function calculate the overlaps of the values data using a bootstrap.
}
\examples{
ch.valuesBootstrapOverlaps (data=valuesData, "tValues", "prompt", params=parameters)
}
\keyword{bootstrap}
\keyword{data}
\keyword{overlap}
\keyword{values}
